requirements.txt
aniso8601==9.0.1
black==24.8.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-RESTful==0.3.10
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
schedule==1.2.2
six==1.16.0
telnyx==2.1.3
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
